DBZ-2092 Modularization; test documentation
