--- debezium-assembly-descriptors/src/main/resources/assemblies/server-distribution.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0"?>
-<assembly xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2" 
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2 http://maven.apache.org/xsd/assembly-1.1.2.xsd">
-  <id>distribution</id>
-  <formats>
-    <format>tar.gz</format>
-    <format>zip</format>
-  </formats>
-  <includeBaseDirectory>false</includeBaseDirectory>
-  <dependencySets>
-      <dependencySet>
-          <outputDirectory>${project.artifactId}/lib</outputDirectory>
-          <unpack>false</unpack>
-          <scope>runtime</scope>
-          <useProjectArtifact>false</useProjectArtifact>
-          <useTransitiveFiltering>true</useTransitiveFiltering>
-          <excludes>
-            <exclude>org.apache.kafka:kafka-tools:*</exclude>
-            <exclude>com.github.luben:zstd-jni:*</exclude>
-            <exclude>org.lz4:lz4-java:*</exclude>
-            <exclude>org.xerial.snappy:snappy-java:*</exclude>
-            <exclude>javax.ws.rs:javax.ws.rs-api:*</exclude>
-            <exclude>org.apache.kafka:connect-file:*</exclude>
-            <exclude>org.glassfish.jersey.*:*:*</exclude>
-            <exclude>org.eclipse.jetty:*:*</exclude>
-            <exclude>org.apache.maven:*:*</exclude>
-          </excludes>
-      </dependencySet>
-  </dependencySets>
-  <fileSets>
-    <fileSet>
-      <!-- Get the files from the top-level directory, which should be above the connectors -->
-      <directory>${project.basedir}/..</directory>
-      <outputDirectory>${project.artifactId}</outputDirectory>
-      <includes>
-        <include>README*</include>
-        <include>CHANGELOG*</include>
-        <include>CONTRIBUTE*</include>
-        <include>COPYRIGHT*</include>
-        <include>LICENSE*</include>
-      </includes>
-      <useDefaultExcludes>true</useDefaultExcludes>
-    </fileSet>
-    <fileSet>
-      <directory>${project.build.directory}</directory>
-      <outputDirectory>${project.artifactId}</outputDirectory>
-      <includes>
-        <include>${project.artifactId}-${project.version}-runner.jar</include>
-      </includes>
-    </fileSet>
-    <fileSet>
-      <directory>${project.build.sourceDirectory}/../distro</directory>
-      <outputDirectory>${project.artifactId}</outputDirectory>
-      <includes>
-        <include>**/*</include>
-      </includes>
-    </fileSet>
-  </fileSets>
-</assembly>
