+++ debezium-server/debezium-server-core/src/main/java/io/debezium/server/BaseChangeConsumer.java
