+++ debezium-server/debezium-server-core/src/test/java/io/debezium/server/DebeziumServerTest.java
