+++ debezium-server/debezium-server-core/src/test/resources/META-INF/services/org.eclipse.microprofile.config.spi.ConfigSource
