+++ debezium-server/debezium-server-core/src/test/resources/log4j.properties
