+++ debezium-server/debezium-server-distro/src/main/resources/distro/run.sh
