+++ debezium-server/debezium-server-kinesis/src/test/resources/META-INF/services/org.eclipse.microprofile.config.spi.ConfigSource
@@ -1,0 +1,1 @@
+io.debezium.server.kinesis.KinesisTestConfigSource
