+++ debezium-server/debezium-server-pubsub/src/test/resources/META-INF/services/org.eclipse.microprofile.config.spi.ConfigSource
@@ -1,0 +1,1 @@
+io.debezium.server.pubsub.PubSubTestConfigSource
